Move to using database rather than direct yaml files
Several screens take a while to load, and could be helped by using
database access instead.  The database model is currently kept very
generic, just essentially as key-value pairs for each server.  This
is done to keep general flexibility, though in the long term there
are definite arguments to be made for using more structured data.