Download tasks now also cache the data
Download and cache separation is good for testing, but not so good when
using the whenever gem.  It doesn't look like there's a good way to chain
the two together in one cron task that way, so I'm copying the cache lines
into the download tasks as well.  Download will download plus cache, while
cache will just cache.  If we don't need the cache anymore for testing,
eventually eliminate it.